crypto: split keys module
