+++ crypto/src/keys/proof.rs
@@ -1,0 +1,6 @@
+use super::{NullifierPrivateKey, SpendAuth, VerificationKey};
+
+pub struct ProofAuthorizationKey {
+    pub ak: VerificationKey<SpendAuth>,
+    pub nsk: NullifierPrivateKey,
+}
